BarberDao implemented.

+++ src/main/kotlin/rs/etf/snippet/rest/ktor/entities/structures/ExtendedReservationWithBarber.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ExtendedReservationWithBarber(
-    var reservationId: Long,
+    var reservationId: Int,
     var clientEmail: String,
     var barberEmail: String,
     var date: String,
@@ -12,6 +12,6 @@ data class ExtendedReservationWithBarber(
     var endTime: String,
     var status: String,
 
-    var barberId: Long,
+    var barberId: Int,
     var barbershopName: String
 )
