+++ src/main/kotlin/rs/etf/snippet/rest/ktor/entities/structures/ExtendedReservationWithClient.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ExtendedReservationWithClient(
-    var reservationId: Long,
+    var reservationId: Int,
     var clientEmail: String,
     var barberEmail: String,
     var date: String,
@@ -12,7 +12,7 @@ data class ExtendedReservationWithClient(
     var endTime: String,
     var status: String,
 
-    var clientId: Long,
+    var clientId: Int,
     var clientName: String,
     var clientSurname: String
 )
