+++ src/main/kotlin/rs/etf/snippet/rest/ktor/entities/structures/ExtendedReviewWithBarber.kt
@@ -4,13 +4,13 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ExtendedReviewWithBarber(
-    var reviewId: Long,
+    var reviewId: Int,
     var clientEmail: String,
     var barberEmail: String,
     var grade: Int,
     var text: String,
     var date: String,
 
-    var barberId: Long,
+    var barberId: Int,
     var barbershopName: String
 )
