+++ src/main/kotlin/rs/etf/snippet/rest/ktor/entities/structures/ExtendedReviewWithClient.kt
@@ -4,14 +4,14 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ExtendedReviewWithClient(
-    var reviewId: Long,
+    var reviewId: Int,
     var clientEmail: String,
     var barberEmail: String,
     var grade: Int,
     var text: String,
     var date: String,
 
-    var clientId: Long,
+    var clientId: Int,
     var clientName: String,
     var clientSurname: String
 )
